+++ services/serverbackup/model_backup_volume_backups_inner.go
@@ -22,10 +22,9 @@ type BackupVolumeBackupsInner struct {
 	Id                   *string `json:"id,omitempty"`
 	LastRestoredAt       *string `json:"lastRestoredAt,omitempty"`
 	LastRestoredVolumeId *string `json:"lastRestoredVolumeId,omitempty"`
-	// Can be cast to int32 without loss of precision.
-	Size     *int64  `json:"size,omitempty"`
-	Status   *string `json:"status,omitempty"`
-	VolumeId *string `json:"volumeId,omitempty"`
+	Size                 *int64  `json:"size,omitempty"`
+	Status               *string `json:"status,omitempty"`
+	VolumeId             *string `json:"volumeId,omitempty"`
 }
 
 // NewBackupVolumeBackupsInner instantiates a new BackupVolumeBackupsInner object