+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Release (2024-XX-XX)
 
+- `serverbackup`: [v0.3.0](services/serverbackup/CHANGELOG.md#v030-2024-11-20)
+  - **Breaking changes**:
+    - `ListBackups200Response` type renamed to `GetBackupsListResponse`
+    - `GetBackupsListResponse` has been removed
+    - `ErrorResponse` has been removed
+    - Added new method `GetBackupSchedulesResponse`
+    - Added new type `EnableServiceResourcePayload`
 - `serverupdate`: [v0.1.0](services/serverupdate/CHANGELOG.md#v010-2024-11-20)
   - Manage your STACKIT Server Updates: `Update`, `UpdateSchedule`, `BackupProperties`
 - `iaas`: [v0.16.0](services/iaas/CHANGELOG.md#v0160-2024-11-08)