+++ services/serverbackup/CHANGELOG.md
@@ -1,3 +1,12 @@
+## v0.3.0 (2024-11-20)
+
+- **Breaking changes**: 
+  - `ListBackups200Response` type renamed to `GetBackupsListResponse`
+  - `GetBackupsListResponse` has been removed
+  - `ErrorResponse` has been removed
+  - Added new method `GetBackupSchedulesResponse`
+  - Added new type `EnableServiceResourcePayload`
+
 ## v0.2.0 (2024-10-14)
 
 - **Feature:** Add support for nullable models